requirements/requirements.txt
fastapi==0.100.0
uvicorn==0.23.1
sqlalchemy==2.0.18
psycopg2-binary==2.9.7
pydantic
python-dotenv==1.0.0
aiogram==3.20.0
